chore(api): regroup cog
